Add locals to map metadata in throw-data.
Adding the locals to the map as proper data would be bad because it
could lead to very large objects being processed in a generic way that
isn't prepared for large objects (e.g., logging), and might encourage
depending on the names of the locals. But having it in the metadata at
least makes it available for interactive debugging.